Just tried to turn on cryptid and got an error
This adds an extra 'end' which screws with the function.
As the check for gallade is in an 'if' with a return I've treated it as
an early out rather than trying to encapsulate the code with an
if and an else.
Seems to work, I can get to the splash screen again but might need a look over?